% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-scenario.R
\name{print.ssd_scenario}
\alias{print.ssd_scenario}
\title{Print a Scenario}
\usage{
\method{print}{ssd_scenario}(x, ...)
}
\arguments{
\item{x}{An \code{ssd_scenario} object.}

\item{...}{Unused.}
}
\value{
The object invisibly.
}
\description{
Prints a summary of an \code{ssd_scenario} object: source description,
simulation settings, fit parameters and hazard-concentration parameters.
}
