% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{ssd_generate_data}
\alias{ssd_generate_data}
\alias{ssd_generate_data.data.frame}
\alias{ssd_generate_data.fitdists}
\alias{ssd_generate_data.tmbfit}
\alias{ssd_generate_data.character}
\alias{ssd_generate_data.function}
\title{Generate Data for Simulations}
\usage{
ssd_generate_data(x, ...)

\method{ssd_generate_data}{data.frame}(x, ..., replace = FALSE, nrow = 6L, nsims = 100L)

\method{ssd_generate_data}{fitdists}(x, ..., dist = "top", nrow = 6L, nsims = 100L)

\method{ssd_generate_data}{tmbfit}(x, ..., nrow = 6L, nsims = 100L)

\method{ssd_generate_data}{character}(x, ..., pars = list(), nrow = 6L, nsims = 100L)

\method{ssd_generate_data}{`function`}(x, ..., args = list(), nrow = 6L, nsims = 100L)
}
\arguments{
\item{x}{The object to use for generating the data.}

\item{...}{Unused.}

\item{replace}{A flag specifying whether to sample with replacement.}

\item{nrow}{A integer vector of the number of rows in the generated data.}

\item{nsims}{A count of the number of data sets to generate.}

\item{dist}{A string specifying the distribution in ssdtools.}

\item{pars}{A named list of the parameter values.}

\item{args}{A named list of the argument values.}
}
\value{
A tibble of generate datasets.
}
\description{
Generate Data for Simulations
}
\section{Methods (by class)}{
\itemize{
\item \code{ssd_generate_data(data.frame)}: Generate Data using data.frame

\item \code{ssd_generate_data(fitdists)}: Generate Data using fitdists Object

\item \code{ssd_generate_data(tmbfit)}: Generate Data using tmbfit Object

\item \code{ssd_generate_data(character)}: Generate Data using string of distribution

\item \code{ssd_generate_data(`function`)}: Generate Data using function

}}
\examples{
ssd_generate_data(ssddata::ccme_boron, nrow = 5, nsims = 3)

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_generate_data(fit, nrow = 5, nsims = 3)

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_generate_data(fit[[1]], nrow = 5, nsims = 3)

ssd_generate_data("lnorm", nrow = 5, nsims = 3)

ssd_generate_data(ssdtools::ssd_rlnorm, nrow = 5, nsims = 3)

}
