% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{ssd_get_seeds}
\alias{ssd_get_seeds}
\title{Get L'Ecuyer-CMRG Seeds}
\usage{
ssd_get_seeds(seed = NULL, ..., nseeds = 100L, nstreams = 1L, skip = 0L)
}
\arguments{
\item{seed}{An integer of the starting seed or NULL.}

\item{...}{Unused.}

\item{nseeds}{A count of the number of new seeds to generate for each stream.}

\item{nstreams}{A count of the number of streams.}

\item{skip}{A count of number of seeds to skip before generating new seeds for each stream.}
}
\value{
A list of lists of L'Ecuyer-CMRG seeds.
}
\description{
Generates a list of streams of L'Ecuyer-CMRG seeds.
}
\examples{
withr::with_seed(10,
ssd_get_seeds(nseeds = 2, nstreams = 2)
)
withr::with_seed(10,
ssd_get_seeds(nseeds = 1, nstreams = 2, skip = 2)
)
}
