% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{ssd_get_seeds_streams}
\alias{ssd_get_seeds_streams}
\title{Get L'Ecuyer-CMRG Streams Seeds}
\usage{
ssd_get_seeds_streams(
  seed = NULL,
  ...,
  nseeds = 1L,
  nstreams = 1L,
  start_seed = 1L,
  start_stream = 1L
)
}
\arguments{
\item{seed}{An integer of the starting seed or NULL.}

\item{...}{Unused.}

\item{nseeds}{A count of the number of new seeds to get for each stream.}

\item{nstreams}{A count of the number of streams.}

\item{start_seed}{A count of the number of the seed to start getting seeds from.}

\item{start_stream}{A count of the number of the stream to start getting streams from.}
}
\value{
A list of lists of L'Ecuyer-CMRG seeds.
}
\description{
Gets a list of streams of L'Ecuyer-CMRG seeds.
}
\examples{
withr::with_seed(10,
ssd_get_seeds_streams(nseeds = 2, nstreams = 2)
)
withr::with_seed(10,
ssd_get_seeds_streams(nseeds = 1, nstreams = 2, start_seed= 2)
)
}
