% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-sims.R
\name{ssd_hc_sims}
\alias{ssd_hc_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_hc_sims(sims, proportion = 0.05, ..., seed = NULL, save_to = NULL)
}
\arguments{
\item{sims}{A list of simulated data.}

\item{proportion}{A numeric vector of proportion values to estimate hazard concentrations for.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_fit_dists()}.}

\item{seed}{An integer of the starting seed or NULL.}

\item{save_to}{NULL or a string specifying a directory to save where the bootstrap datasets and parameter estimates (when successfully converged) to.}
}
\value{
The sims tibble with a list column fits of fistdist objects.
}
\description{
Fit SSD Distributions to Simulated Data
}
