% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scenario.R
\name{ssd_run_scenario}
\alias{ssd_run_scenario}
\alias{ssd_run_scenario.data.frame}
\title{Run Scenario}
\usage{
ssd_run_scenario(x, ...)

\method{ssd_run_scenario}{data.frame}(
  x,
  ...,
  replace = FALSE,
  nrow = c(5L, 10L),
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L
)
}
\arguments{
\item{x}{The object to use for the scenario.}

\item{...}{Unused.}

\item{replace}{A flag specifying whether to sample with replacement.}

\item{nrow}{A count of the number of rows in the generated data which must be between 5 and 1,000,}

\item{seed}{An integer of the starting seed or NULL.}

\item{nsim}{A count of the number of data sets to generate.}

\item{stream}{A count of the stream number.}

\item{start_sim}{A count of the number of the simulation to start from.}
}
\value{
A tibble of nested data sets.
}
\description{
Run Scenario
}
\section{Methods (by class)}{
\itemize{
\item \code{ssd_run_scenario(data.frame)}: Run from data.frame

}}
\examples{
ssd_run_scenario(ssddata::ccme_boron, nsim = 2)

}
