% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{ssd_simulate_data}
\alias{ssd_simulate_data}
\alias{ssd_simulate_data.data.frame}
\alias{ssd_simulate_data.fitdists}
\alias{ssd_simulate_data.tmbfit}
\alias{ssd_simulate_data.character}
\alias{ssd_simulate_data.function}
\title{Generate Data for Simulations}
\usage{
ssd_simulate_data(x, ...)

\method{ssd_simulate_data}{data.frame}(
  x,
  ...,
  replace = FALSE,
  nrow = 6L,
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L,
  .progress = FALSE
)

\method{ssd_simulate_data}{fitdists}(
  x,
  ...,
  dist = "top",
  nrow = 6L,
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L,
  .progress = FALSE
)

\method{ssd_simulate_data}{tmbfit}(
  x,
  ...,
  nrow = 6L,
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L,
  .progress = FALSE
)

\method{ssd_simulate_data}{character}(
  x,
  ...,
  pars = list(),
  nrow = 6L,
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L,
  .progress = FALSE
)

\method{ssd_simulate_data}{`function`}(
  x,
  ...,
  args = list(),
  nrow = 6L,
  seed = NULL,
  nsim = 100L,
  stream = getOption("ssdsims.stream", 1L),
  start_sim = 1L,
  .progress = FALSE
)
}
\arguments{
\item{x}{The object to use for generating the data.}

\item{...}{Unused.}

\item{replace}{A flag specifying whether to sample with replacement.}

\item{nrow}{A count of the number of rows in the generated data which must be between 5 and 1,000,}

\item{seed}{An integer of the starting seed or NULL.}

\item{nsim}{A count of the number of data sets to generate.}

\item{stream}{A count of the stream number.}

\item{start_sim}{A count of the number of the simulation to start from.}

\item{.progress}{Whether to show a \verb{purrr::progress bar}.}

\item{dist}{A string specifying the distribution in the fitdists object or
\code{"top"} to use the distribution with most weight or \code{"multi"} to treat
the distributions as a single distribution.}

\item{pars}{A named list of the parameter values.}

\item{args}{A named list of the argument values.}
}
\value{
A tibble of nested data sets.
}
\description{
A family of functions to generate a tibble of nested data sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{ssd_simulate_data(data.frame)}: Generate data from data.frame

\item \code{ssd_simulate_data(fitdists)}: Generate data from fitdists object

\item \code{ssd_simulate_data(tmbfit)}: Generate data from tmbfit object

\item \code{ssd_simulate_data(character)}: Generate data using distribution name

\item \code{ssd_simulate_data(`function`)}: Generate data using function to generate sequence of random numbers

}}
\examples{
ssd_simulate_data(ssddata::ccme_boron, nrow = 5, nsim = 3)

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_simulate_data(fit, nrow = 5, nsim = 3)

fit <- ssdtools::ssd_fit_dists(ssddata::ccme_boron)
ssd_simulate_data(fit[[1]], nrow = 5, nsim = 3)

ssd_simulate_data("rnorm", nrow = 5, nsim = 3)

ssd_simulate_data(ssdtools::ssd_rlnorm, nrow = 5, nsim = 3)

}
