% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdsims-package.R
\docType{package}
\name{ssdsims-package}
\alias{ssdsims}
\alias{ssdsims-package}
\title{ssdsims: Simulation Analyses for Species Sensitivity Distributions}
\description{
Runs reproducible simulation studies for species sensitivity distribution (SSD) models built on the 'ssdtools' package. Expands a declarative scenario into per-step task tables, draws data, fits distributions, and estimates hazard concentrations, with a 'targets'-based, Hive-partitioned shard pipeline for running studies in parallel or on a cluster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/ssdsims/}
  \item Report bugs at \url{https://github.com/poissonconsulting/ssdsims/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au} (\href{https://orcid.org/0000-0001-5148-6731}{ORCID})
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

}
\keyword{internal}
