% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecuyer-cmrg-seed.R
\name{local_lecuyer_cmrg_seed}
\alias{local_lecuyer_cmrg_seed}
\title{Local L'Euyer-CMRG Seed}
\usage{
local_lecuyer_cmrg_seed(seed, .local_envir = parent.frame())
}
\arguments{
\item{seed}{\verb{[integer(1)]}\cr The random seed to use to evaluate the code.}

\item{.local_envir}{\verb{[environment]}\cr The environment to use for scoping.}
}
\description{
Local L'Euyer-CMRG Seed
}
\examples{

local_lecuyer_cmrg_seed(42)
runif(3)
}
\seealso{
\code{\link[withr:with_seed]{withr::local_seed()}}
}
