% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions for Package}
\arguments{
\item{...}{Unused.}

\item{args}{A named list of the argument values.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by bootstrapping).}

\item{dist}{A string specifying the distribution in the fitdists object or
\code{"top"} to use the distribution with most weight or \code{"multi"} to treat
the distributions as a single distribution.}

\item{nrow}{A count of the number of rows in the generated data which must be between 5 and 1,000,}

\item{nsim}{A count of the number of data sets to generate.}

\item{pars}{A named list of the parameter values.}

\item{replace}{A flag specifying whether to sample with replacement.}

\item{seed}{An integer of the starting seed or NULL.}

\item{start_sim}{A count of the number of the simulation to start from.}

\item{stream}{A count of the stream number.}

\item{.progress}{Whether to show a \verb{purrr::progress bar}.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.
}
\keyword{internal}
