% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions for Package}
\arguments{
\item{...}{Unused.}

\item{nseeds}{A count of the number of new seeds to get for each stream.}

\item{nstreams}{A count of the number of streams.}

\item{seed}{An integer of the starting seed or NULL.}

\item{start_seeds}{A count of the number of the seed to start get new seeds from.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.
}
\keyword{internal}
