% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{print.ssdsims_scenario}
\alias{print.ssdsims_scenario}
\title{Print a Simulation Scenario}
\usage{
\method{print}{ssdsims_scenario}(x, ...)
}
\arguments{
\item{x}{An \code{ssdsims_scenario} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulation Scenario
}
