% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{sdd_get_streams_seeds}
\alias{sdd_get_streams_seeds}
\title{Get L'Ecuyer-CMRG Seeds}
\usage{
sdd_get_streams_seeds(
  seed = NULL,
  ...,
  nseeds = 100L,
  nstreams = 1L,
  start_seeds = 0L
)
}
\arguments{
\item{seed}{An integer of the starting seed or NULL.}

\item{...}{Unused.}

\item{nseeds}{A count of the number of new seeds to generate for each stream.}

\item{nstreams}{A count of the number of streams.}

\item{start_seeds}{A count of number of seeds to start_seedsbefore generating new seeds for each stream.}
}
\value{
A list of lists of L'Ecuyer-CMRG seeds.
}
\description{
Generates a list of streams of L'Ecuyer-CMRG seeds.
}
\examples{
withr::with_seed(10,
sdd_get_streams_seeds(nseeds = 2, nstreams = 2)
)
withr::with_seed(10,
sdd_get_streams_seeds(nseeds = 1, nstreams = 2, start_seeds= 2)
)
}
