% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dists-sims.R
\name{ssd_fit_dists_sims}
\alias{ssd_fit_dists_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_fit_dists_sims(
  datas,
  dists = ssdtools::ssd_dists_bcanz(),
  ...,
  silent = TRUE,
  save_to = NULL,
  return_items = TRUE,
  seed = NULL,
  stream = 1L,
  start_seed = 1L
)
}
\arguments{
\item{datas}{A list of simulated data.}

\item{dists}{A character vector of the distribution names.}

\item{...}{Unused.}

\item{silent}{A flag indicating whether fits should fail silently.}

\item{save_to}{A string of the existing directory to save objects to or NULL.}

\item{return_items}{A flag specifying whether to return a list of the generated items.}

\item{seed}{An integer of the starting seed or NULL.}

\item{stream}{A count of the stream number.}

\item{start_seed}{A count of the number of the seed to start getting seeds from.}
}
\value{
A list of fitdists objects unless \code{return_items = FALSE}.
}
\description{
Fit SSD Distributions to Simulated Data
}
