% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dists-sims.R
\name{ssd_fit_dists_sims}
\alias{ssd_fit_dists_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_fit_dists_sims(
  x,
  dists = ssdtools::ssd_dists_bcanz(),
  ...,
  rescale = FALSE,
  seed = NULL,
  silent = TRUE,
  .progress = FALSE
)
}
\arguments{
\item{x}{A data frame with sim and stream integer columns and a list column of the data frames to fit distributions to.}

\item{dists}{A character vector of the distribution names.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_fit_dists()}.}

\item{rescale}{A flag specifying whether to leave the values unchanged (FALSE) or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values (TRUE) or a string specifying whether to leave the values unchanged ("no") or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values ("geomean") or to logistically transform ("odds").}

\item{seed}{An integer of the starting seed or NULL.}

\item{silent}{A flag indicating whether fits should fail silently.}

\item{.progress}{Whether to show a \verb{purrr::progress bar}.}
}
\value{
The x tibble with a list column fits of fistdist objects.
}
\description{
Fit SSD Distributions to Simulated Data
}
