% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dists-sims.R
\name{ssd_fit_dists_sims}
\alias{ssd_fit_dists_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_fit_dists_sims(
  sims,
  dists = ssdtools::ssd_dists_bcanz(),
  ...,
  silent = TRUE,
  save_to = NULL,
  seed = NULL,
  stream = 1L,
  start_sim = 1L
)
}
\arguments{
\item{sims}{A list of simulated data.}

\item{dists}{A character vector of the distribution names.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_fit_dists()}.}

\item{silent}{A flag indicating whether fits should fail silently.}

\item{save_to}{A string of the existing directory to save objects to or NULL.}

\item{seed}{An integer of the starting seed or NULL.}

\item{stream}{A count of the stream number.}

\item{start_sim}{A count of the number of the simulation to start from.}
}
\value{
The sims tibble with a list column fits of fistdist objects.
}
\description{
Fit SSD Distributions to Simulated Data
}
