% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dists-sims.R
\name{ssd_fit_dists_sims}
\alias{ssd_fit_dists_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_fit_dists_sims(
  x,
  dists = ssdtools::ssd_dists_bcanz(),
  ...,
  seed = NULL,
  silent = TRUE,
  save_to = NULL
)
}
\arguments{
\item{x}{A data frame with sim and stream integer columns and a list column of the data frames to fit distributions to.}

\item{dists}{A character vector of the distribution names.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_fit_dists()}.}

\item{seed}{An integer of the starting seed or NULL.}

\item{silent}{A flag indicating whether fits should fail silently.}

\item{save_to}{A string of the existing directory to save objects to or NULL.}
}
\value{
The x tibble with a list column fits of fistdist objects.
}
\description{
Fit SSD Distributions to Simulated Data
}
