% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-sims.R
\name{ssd_hc_sims}
\alias{ssd_hc_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_hc_sims(
  x,
  proportion = 0.05,
  ...,
  ci = FALSE,
  seed = NULL,
  save_to = NULL,
  .progress = FALSE
)
}
\arguments{
\item{x}{A data frame with sim and stream integer columns and a list column of fitdists objects.}

\item{proportion}{A numeric vector of proportion values to estimate hazard concentrations for.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_hc()}.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by bootstrapping).}

\item{seed}{An integer of the starting seed or NULL.}

\item{save_to}{NULL or a string specifying a directory to save where the bootstrap datasets and parameter estimates (when successfully converged) to.}

\item{.progress}{Whether to show a \verb{purrr::progress bar}.}
}
\value{
The x tibble with a list column hc of data frames produced by applying ssd_hc() to fits.
}
\description{
Fit SSD Distributions to Simulated Data
}
