% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-sims.R
\name{ssd_hc_sims}
\alias{ssd_hc_sims}
\title{Fit SSD Distributions to Simulated Data}
\usage{
ssd_hc_sims(
  x,
  proportion = 0.05,
  ...,
  ci = FALSE,
  nboot = 1000,
  est_method = "multi",
  ci_method = "weighted_samples",
  parametric = TRUE,
  seed = NULL,
  save_to = NULL,
  .progress = FALSE
)
}
\arguments{
\item{x}{A data frame with sim and stream integer columns and a list column of fitdists objects.}

\item{proportion}{A numeric vector of proportion values to estimate hazard concentrations for.}

\item{...}{Additional arguments passed to \code{ssdtools::ssd_hc()}.}

\item{ci}{A flag specifying whether to estimate confidence intervals (by bootstrapping).}

\item{nboot}{A count of the number of bootstrap samples to use to estimate the confidence limits. A value of 10,000 is recommended for official guidelines.}

\item{est_method}{A string specifying whether to estimate directly from
the model-averaged cumulative distribution function (\code{est_method = 'multi'}) or
to take the arithmetic mean of the estimates from the
individual cumulative distribution functions weighted
by the AICc derived weights  (\code{est_method = 'arithmetic'}) or
or to use the geometric mean instead (\code{est_method = 'geometric'}).}

\item{ci_method}{A string specifying which method to use for estimating
the standard error and confidence limits from the bootstrap samples.
Possible values include \code{ci_method = "multi_fixed"} and  \code{ci_method = "multi_free"}
which generate the bootstrap samples using the model-averaged cumulative distribution function
but differ in whether the model weights are fixed at the values for the original dataset
or re-estimated for each bootstrap sample dataset.
The value \code{ci_method = "weighted_samples"} takes bootstrap samples
from each distribution proportional to its AICc based weights and
calculates the confidence limits (and SE) from this single set.
The value \code{ci_method = "MACL"} (was \code{ci_method = "weighted_arithmetic"} but
has been soft-deprecated) which is only included for
historical reasons takes the weighted arithmetic mean of the confidence
limits and \code{ci_method = MGCL} which was included for a research paper
takes the weighted geometric mean of the confidence limits.
The values \code{ci_method = "MAW1"} and \code{ci_method = "MAW2"}
use the two alternative equations of Burnham and Anderson to
model average the weighted standard errors and then calculate the confidence
limits using the Wald approach.
Finally \code{ci_method = "arithmetic"} and \code{ci_method = "geometric"}
take the weighted arithmetic or geometric mean of the values for
each bootstrap iteration across all the distributions and then
calculate the confidence limits (and SE) from the single set of samples.}

\item{parametric}{A flag specifying whether to perform parametric bootstrapping as opposed to non-parametrically resampling the original data with replacement.}

\item{seed}{An integer of the starting seed or NULL.}

\item{save_to}{NULL or a string specifying a directory to save where the bootstrap datasets and parameter estimates (when successfully converged) to.}

\item{.progress}{Whether to show a \verb{purrr::progress bar}.}
}
\value{
The x tibble with a list column hc of data frames produced by applying ssd_hc() to fits.
}
\description{
Fit SSD Distributions to Simulated Data
}
