% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdsims-package.R
\docType{package}
\name{ssdsims-package}
\alias{ssdsims}
\alias{ssdsims-package}
\title{ssdsims: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/ssdsims/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au}
}

}
\keyword{internal}
