% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecuyer-cmrg-seed.R
\name{with_lecuyer_cmrg_seed}
\alias{with_lecuyer_cmrg_seed}
\title{With L'Euyer-CMRG Seed}
\usage{
with_lecuyer_cmrg_seed(seed, code)
}
\arguments{
\item{seed}{\verb{[integer(1)]}\cr The random seed to use to evaluate the code.}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}
}
\description{
With L'Euyer-CMRG Seed
}
\examples{

with_lecuyer_cmrg_seed(42, {
runif(3)
})
}
\seealso{
\code{\link[withr:with_seed]{withr::with_seed()}}
}
